requirements.txt
# Minimum Django and REST framework version
Django>=2.0
djangorestframework>=3.8.2

# Test requirements
pytest-django==3.4.8
pytest==4.4.1
pytest-cov==2.6.1
flake8==3.7.7
django-test-plus==1.4.0
mock==2.0.0

# App requirements
hashids==1.1.0

# Wheel for PyPI installs
wheel==0.24.0

# MkDocs for documentation previews/deploys
mkdocs==0.16.1
